+++ franka_bringup/launch/mobile_teleop.launch.py
@@ -97,12 +97,14 @@ def generate_launch_description():
     return LaunchDescription(
         [
             DeclareLaunchArgument(
-                'controller_names', description='Name of the controller to be used'
+                'controller_names',
+                default_value="swerve_drive_controller",
+                description='Name of the controller to be used',
             ),
             DeclareLaunchArgument(
                 'robot_config_file',
                 default_value=PathJoinSubstitution(
-                    [FindPackageShare('franka_bringup'), 'config', 'franka.config.yaml']
+                    [FindPackageShare('franka_bringup'), 'config', 'tmr.config.yaml']
                 ),
                 description='Path to the robot configuration file to load',
             ),