+++ franka_gazebo_bringup/launch/gazebo_franka_arm_example_controller.launch.py
@@ -39,6 +39,7 @@ def load_controller(context: LaunchContext, controller_name):
         package='controller_manager',
         executable='spawner',
         arguments=[
+            'joint_state_broadcaster',
             controller_name_str,
             '--controller-manager-timeout', '30',
         ],
@@ -180,18 +181,8 @@ def generate_launch_description():
                      executable='rviz2',
                      name='rviz2',
                      namespace=namespace,
-                     arguments=['--display-config', rviz_file, '-f', 'base_link'],
+                     arguments=['--display-config', rviz_file, '-f', 'world'],
                      condition=IfCondition(rviz))
-
-    joint_state_broadcaster = Node(
-        package='controller_manager',
-        executable='spawner',
-        arguments=[
-            'joint_state_broadcaster',
-            '--controller-manager-timeout', '30',
-        ],
-        output='screen'
-    )
 
     launch_controller = OpaqueFunction(
         function=load_controller,
@@ -214,23 +205,7 @@ def generate_launch_description():
         RegisterEventHandler(
             event_handler=OnProcessExit(
                 target_action=spawn,
-                on_exit=[joint_state_broadcaster],
-            )
-        ),
-        RegisterEventHandler(
-            event_handler=OnProcessExit(
-                target_action=joint_state_broadcaster,
                 on_exit=[launch_controller],
             )
         ),
-        RegisterEventHandler(
-            OnShutdown(
-                on_shutdown=[
-                    ExecuteProcess(
-                        cmd=['pkill', '-SIGINT', '-f', 'gz sim'],
-                        name='gz_sim_graceful_shutdown',
-                    )
-                ]
-            )
-        )
     ])