+++ franka_gazebo_bringup/test/test_gazebo_franka_arm_example_controller.py
@@ -24,11 +24,12 @@ import launch_ros.substitutions
 import launch_testing
 import launch_testing.actions
 import rclpy
+import subprocess
 
-TEST_DURATION = 2.0  # sec
+TEST_DURATION = 5.0  # sec
 ROBOT_TYPES = [
-    "fr3", 
-    # "fr3v2", 
+    "fr3",
+    # "fr3v2",
     # "fr3v2_1"
 ]
 CONTROLLERS = [
@@ -38,7 +39,15 @@ CONTROLLERS = [
     "joint_velocity_example_controller",
 ]
 
-params = [(robot_type, controller) for robot_type in ROBOT_TYPES for controller in CONTROLLERS]
+params = [(robot_type, controller)
+          for robot_type in ROBOT_TYPES for controller in CONTROLLERS]
+
+
+def ensure_gz_sim_not_running():
+    # Kill any remaining Gazebo processes
+    # See https://github.com/ros2/launch/issues/545 for details
+    shell_cmd = ['pkill', '-9', '-f', '^gz sim']
+    subprocess.run(shell_cmd, check=False)
 
 
 @launch_testing.parametrize('robot_type, controller', params)
@@ -72,7 +81,8 @@ def generate_test_description(robot_type, controller):
             [
                 launch_description,
                 actions.TimerAction(
-                    period=TEST_DURATION, actions=[launch_testing.actions.ReadyToTest()]
+                    period=TEST_DURATION, actions=[
+                        launch_testing.actions.ReadyToTest()]
                 ),
             ],
         ),
@@ -93,6 +103,7 @@ class TestExampleController(unittest.TestCase):
     def tearDownClass(cls):
         """Shutdown the ROS context."""
         rclpy.shutdown()
+        ensure_gz_sim_not_running()
 
     def test_has_no_error(self, proc_output):
         """Check if any error messages have been logged."""